# Speed up initialization