@morganherlocker Made spline initialization 2+ times faster: 

- Switched to typed arrays.
- Replaced Gaussian elimination code with a classic one [from Wikipedia](https://en.wikipedia.org/wiki/Gaussian_elimination).

```
# before
elements in | times .at() | initialization | execution time
dataset     | is used     | time of spline | of all .at()'s
------------|-------------|----------------|---------------
10          |         100 |            0ms |            0ms
100         |        1000 |           10ms |            4ms
1000        |       10000 |         1060ms |           17ms
2000        |       20000 |         7972ms |           45ms
4000        |       40000 |        59199ms |           82ms

# after
elements in | times .at() | initialization | execution time
dataset     | is used     | time of spline | of all .at()'s
------------|-------------|----------------|---------------
10          |         100 |            0ms |            0ms
100         |        1000 |           16ms |            4ms
1000        |       10000 |          411ms |           18ms
2000        |       20000 |         3585ms |           51ms
4000        |       40000 |        30000ms |           90ms
```