@morganherlocker 2019-12-12
🎉 Thanks @mourner!

Results from my 2017 2.3GHz macbook pro:

```
# before
elements in | times .at() | initialization | execution time
dataset     | is used     | time of spline | of all .at()'s
------------|-------------|----------------|---------------
10          |         100 |            0ms |            0ms
100         |        1000 |            9ms |            3ms
1000        |       10000 |          981ms |           19ms
2000        |       20000 |         7674ms |           47ms
4000        |       40000 |        62091ms |           80ms

# after
elements in | times .at() | initialization | execution time
dataset     | is used     | time of spline | of all .at()'s
------------|-------------|----------------|---------------
10          |         100 |            0ms |            0ms
100         |        1000 |            8ms |            3ms
1000        |       10000 |          423ms |           19ms
2000        |       20000 |         3304ms |           47ms
4000        |       40000 |        28138ms |           78ms
```

I'm going to merge, then publish this as a patch release.